Ticket #286 - textarea now syncs when file is changed
This also changes the behavior of the create event. Before create would
fire _before_ the file was created meaning if you asked for the file
inside of create it would give you undefined. I've moved the create
event into save, rather than importFile, and it fires after the file is
created.