Fix whitespace in editor
Currently, pasting external content with sequences of whitespace into
the editor results in the whitespaces being collapsed into a single one
(which is the default behaviour for HTML elements). Instead we really
want all the whitespaces to be preserved and displayed correctly.

Note that this only applies when pasting content, since pressing the
space bar on the keyboard will produce `&nbsp;` which are not collapsed
in any case.