Added 'main' to point to build JS.
Resolves #387
--

When using NPM as the method to install EpicEditor, and then building the application (using WebPack for example), the main entry point JS file is not resolved.

Ie. These don't work.
```
var EpicEditor = require('epiceditor');
import EpicEditor from 'epiceditor';
```

The `package.json` is missing the `main` declaration that tells the builder which is the entrypoint file.